quack._fetch_modules: [refactor] create `Module` container class
This commit refactors the concept of a `module` as specified in config
into a container class, which also comes with a staticmethod factory
function for creating a `Module` from the configuration. This includes
factoring out the main loop over the various modules into an interator
function. The net result is that the configuration for each module is
accessed by member variables of the instance rather than doing ugly
indexing and key-lookups from the raw config whenever they need to be
referred to.